requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
tensorflow==2.16.1
streamlit==1.31.1
python-dotenv==1.0.0
google-generativeai>=0.8.0
plotly==6.0.1
seaborn==0.13.2